consolidate incident types and add endpoint
